Split unit tests into modules by function
