+++ src/libraries/Wall/src/WallTest.vcxproj.filters
@@ -42,6 +42,21 @@
     <ClCompile Include="..\..\MockDevices\rgb_lcd.cpp">
       <Filter>Resource Files</Filter>
     </ClCompile>
+    <ClCompile Include="CircuitTests.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="IndicatorTests.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="InitTests.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="InputTests.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="OutputTests.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
